requirements.txt
praw
openai==0.28.0
python-dotenv
